Feature/remove jquery (#229)
* remove jquery

+++ spirit/core/static/spirit/scripts/js/store.no-min.js
@@ -1,72 +1,86 @@
-(function() {
-  var $, Storage,
-    bind = function(fn, me){ return function(){ return fn.apply(me, arguments); }; };
 
-  $ = jQuery;
+/*
+    A storage for form elements (inputs, text-areas, ets)
+    It auto-clears on form submission. Also auto-sync across tabs.
+ */
+
+(function() {
+  var Storage,
+    bind = function(fn, me){ return function(){ return fn.apply(me, arguments); }; };
 
   Storage = (function() {
     function Storage(el, lsKey) {
       this.clearStorage = bind(this.clearStorage, this);
       this.updateField = bind(this.updateField, this);
       this.updateStorage = bind(this.updateStorage, this);
-      this.el = $(el);
+      this._updateStorage = bind(this._updateStorage, this);
+      this.el = el;
       this.lsKey = lsKey;
+      this.isUpdating = false;
       this.setUp();
     }
 
     Storage.prototype.setUp = function() {
-      var $form;
       if (typeof localStorage === "undefined" || localStorage === null) {
+        console.log('No localStorage support. Bailing out');
         return;
       }
       if (this.lsKey in localStorage) {
         this.updateField();
       }
-      $(window).on('storage', this.updateField);
-      this.el.on('input change propertychange', (function(_this) {
-        return function() {
-          $(window).off('storage', _this.updateField);
-          _this.updateStorage();
-          $(window).on('storage', _this.updateField);
-        };
-      })(this));
-      $form = this.el.closest("form");
-      return $form.on('submit', this.clearStorage);
+      window.addEventListener('storage', this.updateField);
+      this.el.addEventListener('input', this.updateStorage);
+      this.el.addEventListener('change', this.updateStorage);
+      this.el.addEventListener('propertychange', this.updateStorage);
+      return this.el.closest('form').addEventListener('submit', this.clearStorage);
+    };
+
+    Storage.prototype._updateStorage = function() {
+      var err, value;
+      value = this.el.value;
+      try {
+        localStorage.setItem(this.lsKey, value);
+      } catch (error) {
+        err = error;
+        if (localStorage.length > 0) {
+          localStorage.clear();
+        }
+      }
     };
 
     Storage.prototype.updateStorage = function() {
-      var err, value;
-      value = this.el.val();
+      this.isUpdating = true;
       try {
-        localStorage[this.lsKey] = value;
-      } catch (error) {
-        err = error;
-        localStorage.clear();
+        this._updateStorage();
+      } finally {
+        this.isUpdating = false;
       }
     };
 
     Storage.prototype.updateField = function() {
-      this.el.val(localStorage[this.lsKey]);
+      if (this.isUpdating) {
+        return;
+      }
+      this.el.value = localStorage.getItem(this.lsKey);
     };
 
     Storage.prototype.clearStorage = function() {
-      delete localStorage[this.lsKey];
+      this.isUpdating = true;
+      try {
+        localStorage.removeItem(this.lsKey);
+      } finally {
+        this.isUpdating = false;
+      }
     };
 
     return Storage;
 
   })();
 
-  $.fn.extend({
-    store: function(lsKey) {
-      return this.each(function() {
-        if (!$(this).data('plugin_store')) {
-          return $(this).data('plugin_store', new Storage(this, lsKey));
-        }
-      });
-    }
-  });
+  stModules.store = function(elm, lsKey) {
+    return new Storage(elm, lsKey);
+  };
 
-  $.fn.store.Storage = Storage;
+  stModules.Storage = Storage;
 
 }).call(this);
